Fix build on ARM
svn path=/trunk/KDE/kdeplasma-addons/; revision=1019098

+++ applets/mediaplayer/mediaplayer.cpp
@@ -227,10 +227,10 @@ void MediaPlayer::keyPressEvent(QKeyEvent *event)
         PlayPause();
         break;
     case Qt::Key_Up:
-        audio->setVolume(qMin(qreal(1.0), audio->volume() + 0.1));
+        audio->setVolume(qMin(qreal(1.0), (qreal)(audio->volume() + 0.1)));
         break;
     case Qt::Key_Down:
-        audio->setVolume(qMax(qreal(0.0), audio->volume() - 0.1));
+        audio->setVolume(qMax(qreal(0.0), (qreal)(audio->volume() - 0.1)));
         break;
     case Qt::Key_R:
         RaiseLower();
